moved main helper functions to raylib header file
